Separated customer specific information into other layers.
From this point on, all customer specific changes where hidden (moved to
another layer).
Some files are now completely removed from the git history, but the main changes, should
be preserved and still be visible in the git history, in order to be able to still comprehend
most changes.

Further changes of this commit:
- Fetch from original mainline repositories.
- Append changes/modifications with patches.

Signed-off-by: Benedikt Niedermayr <Benedikt.Niedermayr@mixed-mode.de>